Move simplify_basic_index to utils.shapes
